+++ src/KernelCompiler.cpp
@@ -44,7 +44,7 @@ KernelCompiler::KernelCompiler(const Args& args, const Context* context, const s
   auto& clNames = getClFileNames();
   auto& clFiles = getClFiles();
   assert(clNames.size() == clFiles.size());
-  int n = clNames.size();
+  int n = int(clNames.size());
   for (int i = 0; i < n; ++i) {
     auto &src = clFiles[i];
     files.push_back({clNames[i], src});
@@ -67,12 +67,12 @@ Program KernelCompiler::compile(const string& fileName, const string& extraArgs)
   }
 #ifdef CUDA_BACKEND
   int err = clCompileProgram(p1.get(), 1, &deviceId, args.c_str(),
-                             clSources.size(), (const cl_program*) (clSources.data()), getClFileNames().data(),
+                             u32(clSources.size()), (const cl_program*) (clSources.data()), getClFileNames().data(),
                              nullptr, nullptr);
 #else
   // Skip first file (opencl_compat.cuh) if this is a standard openCL application rather than a CUDA translation
   int err = clCompileProgram(p1.get(), 1, &deviceId, args.c_str(),
-                             clSources.size()-1, (const cl_program*) (clSources.data()+1), getClFileNames().data()+1,
+                             u32(clSources.size()-1), (const cl_program*) (clSources.data()+1), getClFileNames().data()+1,
                              nullptr, nullptr);
 #endif
   if (string mes = getBuildLog(p1.get(), deviceId); !mes.empty()) { log("%s\n", mes.c_str()); }
